lambda/requirements.txt
requests
python-decouple
pymongo
boto3